requirements.txt
streamlit
PyPDF2
pdf2image
Pillow
sentence-transformers==4.1.0
transformers==4.41.0
torch>=2.0.0
tensorflow==2.16.1
tf-keras
protobuf>=3.20.3,<4.21.0
scikit-learn
tqdm
numpy
scipy
requests
huggingface-hub>=0.20.0
google-generativeai
python-dotenv

